==> Genre_Model/1 Pyspark data manipulation.ipynb <==
# Import dependencies
import findspark
from pyspark import SparkFiles
from pyspark.sql import SparkSession
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import accuracy_score
import pandas as pd
import tensorflow as tf
import pandas as pd
from sklearn.preprocessing import MinMaxScaler

# Initialize findspark
findspark.init()

# Initialize the spark session
spark = SparkSession.builder.appName("SK_model").getOrCreate()

# Load in the pre-processed data
df = spark.read.csv("../Resources/filtered_encoded_genre_unbinned.csv", sep=",", header=True, inferSchema=True)
df.show()

# Remove columns from the dataset
df = df.drop("primary_artist", "track_name","artists", "track_id", "album_name")
df.show()

# Convert the data to a pandas Dataframe
p_df = df.toPandas()
p_df.head()

# Remove the specified columns
columns_to_drop = ["num_artists_binned_1", "num_artists_binned_2", "num_artists_binned_3","num_artists_binned_4", "num_artists_binned_5", "num_artists_binned_6"]
p_df_drop = p_df.drop(columns=columns_to_drop)

# Print the updated DataFrame
p_df_drop.head()

# List out columns
columns_list = p_df_drop.columns.tolist()
print(columns_list)

# Create a scaler object
scaler = MinMaxScaler(feature_range=(0, 1))

# Scale 'popularity'
popularity_values = p_df_drop['popularity'].values.reshape(-1, 1)
scaled_popularity = scaler.fit_transform(popularity_values)
p_df_drop['popularity'] = scaled_popularity

# Scale 'loudness'
loudness_values = p_df_drop['loudness'].values.reshape(-1, 1)
scaled_loudness = scaler.fit_transform(loudness_values)
p_df_drop['loudness'] = scaled_loudness

# Scale 'tempo'
tempo_values = p_df_drop['tempo'].values.reshape(-1, 1)
scaled_tempo = scaler.fit_transform(tempo_values)
p_df_drop['tempo'] = scaled_tempo

# Update the columns in the DataFrame with the scaled values
p_df_drop['popularity'] = scaled_popularity
p_df_drop

# Get unique words in 'track_genre' column
unique_genres = p_df_drop['track_genre'].unique()

# Display the unique genres
print(unique_genres)

# Get the count of unique genres
unique_genre_count = p_df_drop['track_genre'].nunique()

# Print the sum of unique genres
print("Sum of unique genres:", unique_genre_count)

# List of unique genres
genres = ['acoustic', 'afrobeat', 'alt-rock', 'alternative', 'ambient', 'anime', 'black-metal', 'bluegrass', 'blues',
          'brazil', 'breakbeat', 'british', 'cantopop', 'chicago-house', 'children', 'chill', 'classical', 'club',
          'comedy', 'country', 'dance', 'dancehall', 'death-metal', 'deep-house', 'detroit-techno', 'disco', 'disney',
          'drum-and-bass', 'dub', 'dubstep', 'edm', 'electro', 'electronic', 'emo', 'folk', 'forro', 'french', 'funk',
          'garage', 'german', 'gospel', 'goth', 'grindcore', 'groove', 'grunge', 'guitar', 'happy', 'hard-rock',
          'hardcore', 'hardstyle', 'heavy-metal', 'hip-hop', 'honky-tonk', 'house', 'idm', 'indian', 'indie-pop',
          'indie', 'industrial', 'iranian', 'j-dance', 'j-idol', 'j-pop', 'j-rock', 'jazz', 'k-pop', 'kids', 'latin',
          'latino', 'malay', 'mandopop', 'metal', 'metalcore', 'minimal-techno', 'mpb', 'new-age', 'opera', 'pagode',
          'party', 'piano', 'pop-film', 'pop', 'power-pop', 'progressive-house', 'psych-rock', 'punk-rock', 'punk',
          'r-n-b', 'reggae', 'reggaeton', 'rock-n-roll', 'rock', 'rockabilly', 'romance', 'sad', 'salsa', 'samba',
          'sertanejo', 'show-tunes', 'singer-songwriter', 'ska', 'sleep', 'songwriter', 'soul', 'spanish', 'study',
          'swedish', 'synth-pop', 'tango', 'techno', 'trance', 'trip-hop', 'turkish', 'world-music']

# Create a dictionary to map genres to numbers
genre_mapping = {genre: i for i, genre in enumerate(genres)}

# Print the genre mapping
for genre, code in genre_mapping.items():
    print(genre, code)

# Convert the genre mapping dictionary to a DataFrame
genre_mapping_df = pd.DataFrame(list(genre_mapping.items()), columns=['genre', 'label'])

# Save the genre mapping DataFrame as a CSV file
genre_mapping_df.to_csv('genre_mapping.csv', index=False)

# Add a new column 'genre_code' to the DataFrame and populate it with the genre codes
p_df_drop['genre_code'] = p_df_drop['track_genre'].map(genre_mapping)
p_df_drop

p_df_drop.to_csv('scaled_data.csv', index=False)

# Specify the columns to drop
columns_to_drop = ["track_genre"]

# Remove the specified columns
p_df_drop_genre = p_df_drop.drop(columns=columns_to_drop)

# Print the updated DataFrame
p_df_drop_genre.head()

p_df_drop_genre.to_csv('dropped_genre_data.csv', index=False)

# List out columns
columns_list = p_df_drop.columns.tolist()
print(columns_list)

Evaluation: After manipulating and exporting a csv file, Logistic regression will be next to test the model
