==> Genre_Model/4 Making predictions.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
from tensorflow.keras.utils import to_categorical
import numpy as np

# Load the dataset
data = pd.read_csv('dropped_genre_data.csv')
data

data.columns

# Separate the features (X) and target (y)
features =data.drop('genre_code', axis=1)# data[['popularity','explicit','danceability','energy','loudness','mode','speechiness']]
target = data[['genre_code']]

print(features.shape)
features.head()

features

print(target.shape)
target.head()

target = np.ravel(target)

features.describe()

# Split the dataset into training and validation/test sets
X_train, X_val, y_train, y_val = train_test_split(features, target, test_size=0.2, random_state=42)

print("Training data dimensions")
print(X_train.shape)
print(y_train.shape)
print()

print("Testing data dimensions")
print(X_val.shape)
print(y_val.shape)


# Define the neural network model

num_inputs = X_train.shape[1]

model = Sequential()
model.add(Dense(50, activation='relu', input_shape=(num_inputs,)))
model.add(Dense(100, activation='relu'))
model.add(Dense(200, activation='relu'))
model.add(Dense(300, activation='relu'))
model.add(Dense(num_classes, activation='softmax'))

# Compile the model
model.compile(loss='sparse_categorical_crossentropy', optimizer='adam', metrics=['accuracy'])

# Train the model
model.fit(X_train, y_train, epochs=15, batch_size=32, validation_data=(X_val, y_val))

model.evaluate(X_val, y_val)

# Evaluate the model's performance
_, accuracy = model.evaluate(X_val, y_val)
print("Validation/Test Accuracy:", accuracy)

# Set a random seed
np.random.seed()

num_of_feats = features.shape[1]

input_feats = []
for i in range(num_of_feats):
    input_feats.append(np.random.rand())

print(input_feats)

# Make predictions using the model
predictions = model.predict(np.array([input_feats]))

# Get the index of the predicted genre code with the highest probability
predicted_genre_index = np.argmax(predictions)

print("Predicted Genre Code:", predicted_genre_index)

import csv

csv_file = "genre_mapping.csv"

# Open the CSV file
with open(csv_file, "r") as file:
    # Create a CSV reader object
    csv_reader = csv.reader(file)

    # Skip the header row (if it exists)
    next(csv_reader)

    # Iterate over each row in the CSV file
    for row in csv_reader:
        genre_code = row[1]
        genre = row[0]

        # Print the genre mapping without labels
        print(genre_code, genre)

Conclusion: Genre type is hard to accurately predict since there may be overlap of similar characteristics for genre types. The data is also skewed and not easy to read. Overall, this was a fun experiment with many iterations.
